+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-0.21.0 - 
+0.21.0 - LogIt Migration
+
+- Add LogIt plugin migration